% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categories.R
\name{AddTodbTable}
\alias{AddTodbTable}
\title{General add to database table function}
\usage{
AddTodbTable(item, dbTable, Id = "id", field = "name", con = .rqda$qdacon)
}
\description{
General add to database table function
}
