% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setting.R
\name{apply_font_css}
\alias{apply_font_css}
\title{Apply current font settings via CSS to a GTK widget
Zoom factor (.rqda$font.zoom) scales the point size.}
\usage{
apply_font_css(widget, family = .rqda$font.family, size = .rqda$font.size)
}
\description{
Apply current font settings via CSS to a GTK widget
Zoom factor (.rqda$font.zoom) scales the point size.
}
\keyword{internal}
