% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Setting.R
\name{apply_startup_settings}
\alias{apply_startup_settings}
\title{Apply dark mode at startup if saved setting requires it}
\usage{
apply_startup_settings()
}
\description{
Apply dark mode at startup if saved setting requires it
}
\keyword{internal}
