% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{AttrNamesUpdate}
\alias{AttrNamesUpdate}
\title{Update attribute names widget}
\usage{
AttrNamesUpdate(Widget = NULL, sortByTime = FALSE, decreasing = FALSE)
}
\description{
Update attribute names widget
}
