% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_gui.R
\name{build_rqda_gui}
\alias{build_rqda_gui}
\title{Build RQDA GUI (called by application activate)}
\usage{
build_rqda_gui(application)
}
\description{
Build RQDA GUI (called by application activate)
}
\keyword{internal}
