% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{ClearMark}
\alias{ClearMark}
\title{Clear marks/highlights from a range}
\usage{
ClearMark(
  textview = NULL,
  min = 0,
  max,
  clear.fore.col = TRUE,
  clear.back.col = TRUE
)
}
\arguments{
\item{textview}{GtkTextView widget}

\item{min}{Start position}

\item{max}{End position}

\item{clear.fore.col}{Clear foreground color}

\item{clear.back.col}{Clear background color}
}
\description{
Clear marks/highlights from a range
}
