% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{.add_attribute_dialog}
\alias{.add_attribute_dialog}
\title{Dialog to add attribute with name and class selection}
\usage{
.add_attribute_dialog()
}
\description{
Dialog to add attribute with name and class selection
}
\keyword{internal}
