% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{.get_selected_attr}
\alias{.get_selected_attr}
\title{Get currently selected attribute name - reads fresh listbox ref from .rqda}
\usage{
.get_selected_attr()
}
\description{
Get currently selected attribute name - reads fresh listbox ref from .rqda
}
\keyword{internal}
