% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesFun.R
\name{.inline_edit}
\alias{.inline_edit}
\title{Helper function to view file
Replace file viewer content with editor in the same window}
\usage{
.inline_edit(FileName, fid, frame, viewer_win, codingTable = "coding")
}
\description{
Helper function to view file
Replace file viewer content with editor in the same window
}
\keyword{internal}
