% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{FileAttrFun}
\alias{FileAttrFun}
\title{Open attribute editor for a file by ID (called from right-click menu)}
\usage{
FileAttrFun(fileId = NULL, title = NULL)
}
\description{
Open attribute editor for a file by ID (called from right-click menu)
}
