% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Attributes.R
\name{FileAttribute_Button}
\alias{FileAttribute_Button}
\title{File Attribute button handler - opens value editor for selected file}
\usage{
FileAttribute_Button(label, container = NULL, FileWidget = NULL, ...)
}
\description{
File Attribute button handler - opens value editor for selected file
}
