% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{gconfirm}
\alias{gconfirm}
\title{Show confirmation dialog}
\usage{
gconfirm(message, title = "Confirm", parent = NULL, icon = "question")
}
\arguments{
\item{message}{Question to ask}

\item{parent}{Parent window (optional)}

\item{icon}{Icon type (usually "question")}
}
\value{
TRUE if yes, FALSE if no
}
\description{
Show confirmation dialog
}
