% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwidgets_compat.R
\name{gfile}
\alias{gfile}
\title{File chooser dialog}
\usage{
gfile(
  text = "Select a file",
  type = c("open", "save", "selectdir"),
  filter = NULL,
  multi = FALSE,
  ...
)
}
\description{
File chooser dialog
}
