% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{gmessage}
\alias{gmessage}
\title{Show message dialog}
\usage{
gmessage(message, icon = "info", parent = NULL, container = FALSE)
}
\arguments{
\item{message}{Message to display}

\item{icon}{Icon type: "info", "warning", "error", "question"}

\item{parent}{Parent window (optional)}
}
\description{
Show message dialog
}
