% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwidgets_compat.R
\name{gnotebook}
\alias{gnotebook}
\title{Create a notebook (tabbed interface)}
\usage{
gnotebook(tab.pos = 3, container = NULL, closebuttons = FALSE, ...)
}
\description{
Create a notebook (tabbed interface)
}
