% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwidgets_compat.R
\name{gtable}
\alias{gtable}
\title{Create a table/list widget}
\usage{
gtable(
  items = character(0),
  container = NULL,
  multiple = FALSE,
  handler = NULL,
  ...
)
}
\description{
Create a table/list widget
}
