% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialogs.R
\name{gtext_window}
\alias{gtext_window}
\title{Create a simple text display window}
\usage{
gtext_window(text, title = "Text", width = 600, height = 400)
}
\arguments{
\item{text}{Text to display}

\item{title}{Window title}

\item{width}{Window width}

\item{height}{Window height}
}
\description{
Create a simple text display window
}
