% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gwidgets_compat.R
\name{gwindow}
\alias{gwindow}
\title{Create a window}
\usage{
gwindow(
  title = "",
  parent = NULL,
  width = 400,
  height = 300,
  visible = TRUE,
  handler = NULL
)
}
\description{
Create a window
}
