% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{HL}
\alias{HL}
\title{Highlight a range with specific colors}
\usage{
HL(textview = NULL, index, fore.col = .rqda$fore.col, back.col = NULL)
}
\arguments{
\item{textview}{GtkTextView widget}

\item{index}{Data frame with columns: start position, end position}

\item{fore.col}{Foreground color}

\item{back.col}{Background color}
}
\description{
Highlight a range with specific colors
}
