% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilesFun.R
\name{ImportFileText}
\alias{ImportFileText}
\title{Import file into project}
\usage{
ImportFileText(
  paths = NULL,
  encoding = .rqda$encoding,
  con = .rqda$qdacon,
  ...
)
}
\description{
Import file into project
}
