% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{LoadCodings}
\alias{LoadCodings}
\title{Load and display existing codings for the current file}
\usage{
LoadCodings(fid, codingTable = "coding")
}
\arguments{
\item{fid}{File ID}

\item{codingTable}{Which coding table}
}
\description{
Load and display existing codings for the current file
}
