% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{mark}
\alias{mark}
\title{Mark selected text with a code}
\usage{
mark(
  textview = NULL,
  fore.col = .rqda$fore.col,
  back.col = .rqda$back.col,
  codingTable = "coding"
)
}
\arguments{
\item{textview}{GtkTextView widget}

\item{fore.col}{Foreground color}

\item{back.col}{Background color}

\item{codingTable}{Which coding table to use}
}
\value{
List with start, end, and text
}
\description{
Mark selected text with a code
}
