% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{markRange}
\alias{markRange}
\title{Mark a specific range (used for loading existing codings)}
\usage{
markRange(
  textview = NULL,
  from,
  to,
  rowid,
  fore.col = .rqda$fore.col,
  back.col = .rqda$back.col,
  codingTable = "coding"
)
}
\arguments{
\item{textview}{GtkTextView widget}

\item{from}{Start position}

\item{to}{End position}

\item{rowid}{Database rowid for this coding}

\item{fore.col}{Foreground color}

\item{back.col}{Background color}
}
\description{
Mark a specific range (used for loading existing codings)
}
