% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProjectFun.R, R/utils.R
\name{rqda_exe}
\alias{rqda_exe}
\title{Execute SQL statement}
\usage{
rqda_exe(sql)

rqda_exe(sql)
}
\arguments{
\item{sql}{sql-text}
}
\description{
Execute SQL statement

Wrapper for dbExecute
}
