% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos_support.R
\name{setup_macos_shortcuts}
\alias{setup_macos_shortcuts}
\title{Setup Cmd key shortcuts for macOS}
\usage{
setup_macos_shortcuts(window)
}
\description{
Setup Cmd key shortcuts for macOS
}
\keyword{internal}
