% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macos_support.R
\name{setup_native_window}
\alias{setup_native_window}
\title{Make window behave like native macOS application}
\usage{
setup_native_window(window)
}
\description{
Make window behave like native macOS application
}
\keyword{internal}
