% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TextSelection.R
\name{sindex}
\alias{sindex}
\title{Get selection information from text widget}
\usage{
sindex(textview = NULL, includeAnchor = TRUE, codingTable = "coding")
}
\arguments{
\item{textview}{GtkTextView widget}

\item{includeAnchor}{Whether to include anchor positions}

\item{codingTable}{Which coding table to use}
}
\value{
List with selection details
}
\description{
Get selection information from text widget
}
