% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categories.R
\name{UpdateCodeofCatWidget}
\alias{UpdateCodeofCatWidget}
\title{Update code of category widget}
\usage{
UpdateCodeofCatWidget(Widget = .rqda$.CodeofCat, con = .rqda$qdacon)
}
\description{
Update code of category widget
}
