% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categories.R
\name{UpdateFileofCatWidget}
\alias{UpdateFileofCatWidget}
\title{Update file of category widget}
\usage{
UpdateFileofCatWidget(Widget = .rqda$.FileofCat, con = .rqda$qdacon)
}
\description{
Update file of category widget
}
