% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Categories.R
\name{UpdateTableWidget}
\alias{UpdateTableWidget}
\title{General update table widget function}
\usage{
UpdateTableWidget(
  Widget,
  FromdbTable,
  con = .rqda$qdacon,
  sortByTime = FALSE,
  decreasing = FALSE
)
}
\description{
General update table widget function
}
